pass secretStorage instead of using global singleton
`secretStorage` is not intended by the VS Code extension API to be a global. It is passed to the `activate` function. I believe this is because VS Code's extension API designers felt it was important to be explicit about which parts of your code have access to the secrets, to avoid mistakes. While there is no specific problem this change fixes, I think it's better to be explicit about the use of secret storage and secret data flow.